+++ src/Command/DeleteStudentCommand.java
@@ -5,25 +5,37 @@ import Model.Person.Student;
 
 public class DeleteStudentCommand implements Command {
     private final StudentDao studentDao;
-    private final Student student;
+    private final String studentId;
+    private Student student;
 
     public DeleteStudentCommand(StudentDao studentDao, Student student) {
         this.studentDao = studentDao;
-        this.student = student;
+        this.studentId = student.getPersonID();
     }
 
     @Override
-    public void execute() {
+    public void execute() throws Exception{
+        // Retrieve the student from the DAO
+        student = studentDao.get(studentId).orElse(null);
+        if (student == null) {
+            throw new Exception("Student not found during delete.");
+        }
         studentDao.delete(student);
     }
 
     @Override
-    public void undo() {
-        studentDao.save(student);
+    public void undo() throws Exception{
+
+        if (student != null) {
+            studentDao.save(student);
+        } else {
+            throw new Exception("No student to undo.");
+        }
+
     }
 
     @Override
     public String getDescription() {
-        return "Delete Student: " + student.getPersonID();
+        return "Delete Student: " + studentId;
     }
 }
